config npm
